+++ src/utils/__tests__/getUnitFromValueNode-test.js
@@ -10,6 +10,8 @@ test("getUnitFromValueNode", t => {
   t.equal(getUnitFromValueNode(valueParser("1PX").nodes[0]), "PX")
   t.equal(getUnitFromValueNode(valueParser("100%").nodes[0]), "%")
   t.equal(getUnitFromValueNode(valueParser("100").nodes[0]), "")
+  t.equal(getUnitFromValueNode(valueParser("#fff").nodes[0]), null)
+  t.equal(getUnitFromValueNode(valueParser("#000").nodes[0]), null)
   t.equal(getUnitFromValueNode(valueParser("\"100\"").nodes[0]), null)
   t.equal(getUnitFromValueNode(valueParser(" ").nodes[0]), null)
   t.equal(getUnitFromValueNode(valueParser("/").nodes[0]), null)