+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Head
+# 6.4.1
 
 - Fixed: `shorthand-property-no-redundant-values` now ignores `background`, `font`, `border`, `border-top`, `border-bottom`, `border-left`, `border-right`, `list-style`, `transition` properties.
 - Fixed: `unit-no-unknown` now ignores hex colors.